+++ src/backend/joanie/core/helpers.py
@@ -2,7 +2,9 @@
 Helpers that can be useful throughout Joanie's core app
 """
 
-from joanie.core import enums
+from django.db.models.query import QuerySet
+
+from joanie.core import enums, models
 from joanie.core.exceptions import CertificateGenerationError
 
 
@@ -12,8 +14,15 @@ def generate_certificates_for_orders(orders):
     then return the count of generated certificates.
     """
     total = 0
+    if isinstance(orders, QuerySet):
+        orders_queryset = orders
+    elif isinstance(orders, list):
+        orders_queryset = models.Order.objects.filter(pk__in=orders)
+    else:
+        raise ValueError("orders must be either List or QuerySet")
+
     orders_filtered = (
-        orders.filter(
+        orders_queryset.filter(
             state=enums.ORDER_STATE_VALIDATED,
             certificate__isnull=True,
             product__type__in=enums.PRODUCT_TYPE_CERTIFICATE_ALLOWED,